requirements.txt
Flask
Flask-CORS
torch
torchvision
torchaudio
git+https://github.com/openai/whisper.git
openai-whisper
pyaudio
